Stop processing command-line options at first non-option argument
This makes it much easier to pass options to the program being
executed by fips. The options can be passed directly, such as:

	fips glxgears -fullscreen

Previous, to this commit, one would have to use a syntax such as:

	fips -- glxgears -fullscreen

to prevent fips from trying to interpret the "-fullscreen" argument as
options to fips itself.