Add collection of (AMD_performance_monitor) performance counters to fips
The implementation involves a linked-list of outstanding
performance-monitor queries next to the existing list of outstanding
timer queries.

The results from the performance counters are stored (without any
interpretation in an array of values next to the existing time values
within each op_metrics_t value for each operation.

The numbers are currently printed with simple counter numbers (no
names and no units) and with the values divided by 1e6. Counters with
values of zero are not printed.

Next steps from here that will make things useful:

1. Use relative number of cycels in each stage to apportion measured
   shader time among the various stages, (so that per-stage time
   numbers are sorted in the final report).

2. Print percentage active, (by looking at per-stage active and stall times)

3. Print names for counters (other than per-stage active and stall
   which will be used in the above two calculations).

4. Fix to silently ignore performance counters if the
   AMD_performance_monitor extension is not available.