Fix conversion from group IDs to group array indices.
The loop that found the array index for a particular group based on the
group ID had a subtle bug: it compared against "i" instead of group_id.

In the i965 implementation, the first group happens to have ID 0, which
meant that the loop would always select the first group (since the ID
equals the array index).  This led to assertion failures about the
number of counters in each group.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>