Fix to still print metrics when there are no per-stage cycle-count metrics
There are at least two cases where fips will have no per-stage cycle counts:

	1. The AMD_performance_monitor extension is not available

	2. The extension is there, but fips cannot find any per-stage
	   cycle counters

In either case, the previous code would print no timing
information. The problem is that fips was using the number of detected
per-stage cycle counters (in num_shader_stages) as the basis for
allocation of the results array. So with this value being 0, nothing
was allocated, nothing was stored, and nothing was printed.

Here, we fix this by instead allocating space for one result per
operation, and ensuring that ll of the measured time is reported for
that operation.