context: Avoid doing excessive work if application re-sets the same context
I've heard rumors that firefox (at least as of some version) had a
tendency to call glXMakeCurrent with the current context on every frame.

As of this commit, fips will notice the current context being set with
MakeCurrent and will do not tear things down and re-create them in
this case.