Push oustanding-counter data down from context.c into metrics.c
Here, we introduce a new metrics_t structure which is responsible for
kepping track of lists of outstanding counters as well as the array of
collected counter results.

Now, context_t is much cleaner, containing only the ID of the system
context, the metric_info_t and the metrics_t.