Allow fips to compile with OpenGL without GLfixed
The Mesa 9.1 releases don't include the GLfixed datatype, so fips was
failing to compile aginst those versions of Mesa, (since it was trying
to add wrappers using the GLfixed datatype).

But since the underlying functions don't exist in libGL.so, fips
doesn't need the wrappers anyway.

In this commit, we key off of the GL_GLEXT_VERSION macro from glext.h
to decide whether or not to include these newer wrappers. This fixes
compilation of fips on Mesa 9.1.