Add fips-find-lib program.
When the target application differs in architecture from the fips
executable, dlopen cannot be used in fips to find the GL/EGL libary
that the target would have linked against.

fips-find-lib compiles for i386 and x86_64, and reports the library
path that will used for linking.  It accepts a single parameter, which
must be 'EGL' or 'GL', indicating which library is requested.

Fips calls the right find-lib program depending on the arch of the
target application.