Handle calls of MakeCurrent when the context has not changed.
When eglMakeCurrent/glXMakeCurrent is called, fips unconditionally
cleans up metrics query resources.  This must be done in
context_leave() before the context changes in the GL.

Unfortunately, when MakeCurrent was called twice with the same
context, the metrics resources were cleaned up, but the parent context
was not recreated in the subsequent context_enter().  This resulted in
a crash when instrumenting the GpuTest FurMark benchmark.

In order to properly detect whether metrics resources should be
cleaned up, the future context id must be checked in context_leave().